ci_evaluator/back/requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
cycler==0.11.0
dill==0.3.4
filelock==3.4.0
Flask==2.0.2
Flask-SQLAlchemy==2.5.1
fonttools==4.28.1
greenlet==1.1.2
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
joblib==1.1.0
kiwisolver==1.3.2
MarkupSafe==2.0.1
matplotlib==3.5.0
nlp==0.4.0
nltk==3.6.5
numpy==1.21.4
packaging==21.2
pandas==1.3.4
Pillow==8.4.0
psycopg2-binary==2.9.2
pyarrow==6.0.0
pyparsing==2.4.7
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
regex==2021.11.10
requests==2.26.0
scikit-learn==1.0.1
scipy==1.7.2
seaborn==0.11.2
setuptools-scm==6.3.2
six==1.16.0
sklearn==0.0
SQLAlchemy==1.4.27
threadpoolctl==3.0.0
tomli==1.2.2
tqdm==4.62.3
urllib3==1.26.7
Werkzeug==2.0.2
xxhash==2.0.2
tablib==3.1.0
tablib[html]==3.1.0
Flask-Table==0.5.0
